requirements.txt
certifi==2018.4.16
chardet==3.0.4
furl==1.0.1
idna==2.6
oauthlib==2.0.7
orderedmultidict==0.7.11
py-trello==0.10.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
